Fix URL that points to wordpress
Before the siteurl was used to determine the URL to use to call the
wordpress REST API. With the docker setup this no longer works because
the request needs to made over the docker network. For this reason the
currently unused setting ckanext.switzerland.wp_url is now used to set
the host URL for wordpress.