feat: Use a custom helper to set the language picker
We now set the CKAN config using ckanext-envvars, but the locale settings are set when CKAN is initialising, before ckanext-envvars can do that. That means the locale settings are those in the default config.

Here, we need to use our own prioritised list of locales, so if we can't get it from the config, store-bought is fine.