fix: Get all datasets, not just the default 10
Previously, this request got only 10 datasets, because 'rows' was not specified. The length of the list of datasets was then assumed to be the number of datasets available, even if the value of datasets.count was higher.

I have set this to 1000, which is the hard upper limit of datasets that CKAN will return on one query, because I did not want to refactor the rest of this Vue code, and because we currently only have 65 datasets, so we are unlikely to run into problems with long response times. #overlyHonestCommitMessages