fix: Load search javascript after jQuery is added
Necessary because the search code relies on jQuery already having been loaded. Normally we can include an asset wherever we like in a template and it's not encouraged to put them in the scripts block unless it's a special case. In this case, it was getting included before jQuery and causing an error.