feat: Handle ckan config keys that contain _ instead of -
This is a change caused by our switch to using ckanext-envvars to define configuration. Env vars cannot contain -, so we substitute with _. Since the S3 bucket names contain -, we need to replace that with _ before we can use the bucket names as keys to find the configuration for the harvester.