tests: Adjust order of unordered resources
We actually only care about the ordering of the two resources that match the resource_regex: 20160902.csv and 20160901.csv. The other two resources are unordered and so long as they are in positions 0 and 1, it doesn't matter which is which.

On updating to CKAN 2.10 and Python 3, it seems that the order they get put in has changed. We don't care, so just update the test.