tests: Make sure ckan.site_url config value is set when needed
When testing in the Docker container, we use the test.ini, which sets the ckan.site_url to the test value. However, locally, this is overridden by the value that is set via ckanext-envvars from the env var CKAN_SITE_URL. Let's make sure it's always set when we run a test that relies on this config value.