tests: Install dev-requirements with '--ignore-installed'
This should prevent installation failure caused by the fact that the version of black we use requires a higher version of packaging than is installed on the test docker container.

In that case, pip tries to uninstall packaging before installing our dev-requirements (including black), but the uninstallation fails because packaging was not originally installed by pip.