fix: Do not re-import old files
Previously, we would remove a file from the list of files to be imported if it not been updated since the last harvester run AND it was already on the dataset as a resource.

In the last step of the harvesting process ('finalizer'), we delete all resources beyond a certain limit, which can be very small (3-50 resources). This means that an old file, not recently updated, may well have been imported once as a resource and then deleted - in this case, we don't want to import it again!

In special cases where we do want to import old files, the harvester config option 'force_all' should be used.