feat: Convert datetimes differently for display and API
The CKAN templates need UTC datetimes with no time zone info. The API wants datetimes in the Europe/Zurich time zone, with time zone info.

For other cases when package_show is called, we don't want to reformat the datetimes at all: it is called a lot internally when updating resources, etc.