chore: Remove old validator not working as intended
swiss_date was set as an output validator for the 'modified' and 'issued' fields on datasets and resources. Since it was called on output, not on creating/updating datasets, the errors it added were never logged anywhere and didn't actually change anything. Furthermore, it was validating that values were dates, when in fact we save 'modified' and 'issued' values as isoformat datetime strings.