feat: Improve resource downloading
We don't need to set the Content-Type and Content-Disposition headers separately like we were, as Flask can do it for us. It's also safe to pass in the mimetype like this, because if there is no mimetype saved for the resource and we pass in None, Flask will guess mimetype based on the file extension.