feat: Translate label of resource field 'media_type' properly
The resource's media_type and format are distinct and it is confusing to label them the same way. Media type is the new name for MIME type, and its value has the format type/subtype, e.g. text/plain. Format is an all-caps string used for search, filtering, and display of datasets and resources, e.g. TXT.