feat: Add resource field 'mimetype_inner' to our schema
This field comes from CKAN. It can be used to show the media_type of files inside a ZIP file, for example: {'media_type': 'application/zip', 'mimetype_inner': 'text/plain'}.

Previously, we were setting this field (with an incorrect value) in the harvester but not using it anywhere else. Since it wasn't in our schema, we could not display it or make it editable in the resource edit page.

CKAN also includes the resource field 'mimetype', which for a ZIP file would be 'application/zip'. This field has the same meaning as media_type (MIME type is the old name for media type) so I am leaving it out of our schema, as we already define the field media_type.