feat: Override ckanext-scheming template to remove broken block
In ckanext-scheming, this template includes a block that was copied over from ckan core (see commit at https://github.com/ckan/ckanext-scheming/commit/e564b9ae9638fdb1e936143307e7296318580c35). This would add several rows to the middle of the resource additional-info table, with the class 'toggle-more' so theycan be expanded and collapsed. However, this toggle does not work when the rows are in the middle of the table, so they are permanently collapsed and break the striping of the table's rows. These rows also contain information that is a duplicate of what we are showing for the resource using the scheming fields.