feat: Don't localise package fields in before_dataset_show
This commit also separates out the before_view handling of groups, organisations and packages, which were all clumped together.

For packages, we want to localise multilingual fields *not* in the before_dataset_show method, but in the templates. This removes the issue that sometimes we get a dataset from Solr, where the custom scheming fields have already been extracted from the package extras, and we can localise it easily in before_dataset_show; but sometimes we get the dataset straight from the db, and the custom fields are still in the extras. In this case, we cannot localise the multilingual fields in before_dataset_show. This means that templates were sometimes getting a single language's field value, and sometimes getting the multilingual dictionary. If we expect to do all localisation in the templates, this is avoided.