fix: Do not run import_stage for repeated harvest object
Sometimes the fetch process gets a harvest object from the queue that has already been processed. This happens after all of a job's harvest objects have been processed, including the one that deletes the temp folder storing files to add as resources. If this harvest object gets passed to the import stage, we get an error that it cannot be processed because the folder does not exist.

To prevent this, the fetch stage should return 'unchanged', which prevents the harvest object being passed on to the import stage.

We have not figured out how these harvest objects get passed on to the fetch process a second time, after they should have been saved as complete and removed from the queue, so this is still just a workaround.