chore: Remove blueprint that didn't do anything
We previously (in CKAN 2.9 and before) had an OgdchResourceController that allowed downloading of deleted resource files. This is no longer possible due to changes in how CKAN keeps old versions of resources in the database, so we removed the functionality, leaving this stub of a blueprint that was not actually used.