refactor: Rename methods and vars for greater clarity
We no longer reduce package data to a single language in the before_dataset_show stage, and we have separated out the before-show handling of packages vs. groups and orgs. This change renames the relevant vars and methods to reflect that. It also separates the _group_or_org_map_ckan_default_fields and _package_map_ckan_default_fields methods, also for clarity.