chore: Remove OgdchResourcePlugin
This plugin now only has one method: _ignore_field, which is called as part of it's parent class OgdchLanguagePlugin's before_view method. However, OgdchResourcePlugin.before_view is never called: the before_view method in the IResourceController interface was renamed to before_resource_view in CKAN 2.10. When we updated to CKAN 2.10, we apparently neglected to create the method OgdchResourcePlugin.before_resource_view, and nothing bad has happened because of this, so it's save to remove this whole plugin.