feat: Adapt helper strxfrm to our needs
This helper converts a value into a form that can be used as a sort key - for example, it converts a group's display_name so that a list of groups can be sorted by display_name. Our group titles are multilingual, so we need to localise the value before this can be done.