feat: Delete extra resources by id, not filename
If we get into a state where we have duplicate resources for the same filename (e.g. because we created a new resource for the filename but deleting the old resource failed), we must not delete resources by filename in the finalize step! If we do, we will delete both the new and the old resource for the filename.

This method is already a lot simpler than it used to be, since CKAN no longer stores lots of old versions of resources in the database, so it's safe to change it to delete resources by id.