feat: Re-add check that dataset exists
The call to get the dataset was removed in ad6e575f559faaf6a58c332fc8cc4fc0ddee4428 because we were no longer using the dataset or package in this block. Catching a NotFound error was thus useless here. But actually, it's useful to check whether the dataset exists, and import all files if it doesn't.