fix: Log a warning instead of returning an error
There are cases where files have been updated at the source, but the
permalink should remain the same. If a file has been edited at the
source, but there is no new file with a later date in the filename, then
we should import the updated file but not update the permalink.

For this reason, we can't return here and save an error. We have to log
a warning instead.