feat: Don't use the Masonry widget to arrange showcases
Masonry (for which media-grid.js is just a wrapper) arranges differently-sized items in the media grid so that they fit. However, we want to display the showcase items as equally-sized tiles, which is handled with some custom CSS defined via the CKAN GUI. We also want to display these items really in the order that the user has selected in the search form, not nearly in that order but adjusted to fit their different sizes (which are now forced to be the same size). So, we don't want to use Masonry.